use cm6 in tutorial + move highlighting to hook + optimize tutorial loading time
